test(functional): make cleanup listener evidence deterministic
